Group HDFS internals by service
