+++ rust/hdfs-native/Cargo.toml
@@ -28,6 +28,7 @@ once_cell = "1"
 prost = "0.14"
 prost-types = "0.14"
 rand = "0.10"
+regex = "1"
 reqwest = { version = "0.13", optional = true, default-features = false, features = ["json"] }
 serde = { version = "1", optional = true, features = ["derive"] }
 serde_json = { version = "1", optional = true }
