+++ rust/hdfs-native/src/encryption/codec.rs
