+++ rust/hdfs-native/src/encryption/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod codec;
+#[cfg(feature = "kms")]
+pub(crate) mod kms;
