--- rust/hdfs-native/src/hdfs/mod.rs
@@ -1,7 +1,0 @@
-pub(crate) mod block_reader;
-pub(crate) mod block_writer;
-pub(crate) mod connection;
-pub(crate) mod crypto;
-pub(crate) mod protocol;
-pub(crate) mod proxy;
-pub(crate) mod replace_datanode;
