--- rust/hdfs-native/src/security/mod.rs
@@ -1,6 +1,0 @@
-#[cfg(feature = "kms")]
-pub(crate) use hadoop_native::security::gssapi;
-#[cfg(feature = "kms")]
-pub(crate) mod kms;
-pub mod sasl;
-pub mod user;
